Add servlet to handle livy client sessions.
This is not complete; it allows creating / tearing down sessions and
submitting synchronous / asynchronous jobs, but there's currently no
way to retrieve a job's results (or even monitoring it's state, since
the backend code for that is not implemented).

Also includes a new base test class for the servlets that does away
with a lot of copy & paste.