Add support to bypass pre-serialized jobs in client-local.
To implement the HTTP-based protocol for Spark client API, the
Livy server needs to be able to bypass data serialized in the
client to the remote driver, without trying to deserialize it
in the process, because it doesn't have the necessary client
classes to do that.

This change adds a private api in client-local and some shared
helper code to implement this bypass; the remote driver will
deserialize these jobs as part of executing them. The code
running inside the Livy server becomes a dumb bypass for the
job and its result.

- Removed "Serializable" from API, since it only adds noise when
  using Kryo.
- Added a "client-common" package with shared serde code to implement
  bypass jobs.
- Added "private" API to client-local to send already serialized jobs
  to the remote context.