Modify bypass jobs to use polling to fetch status.
Right now, bypass jobs work like normal jobs; they return data to the
client as they make progress, and return the results at the end.

When the client in the picture is the Livy server, that's not desirable.
That means that the Livy server would have to store all data generated
by client jobs until the client decides to poll for it; meaning that jobs
could consume lots of memory on the Livy server just by returning a lot
of data.

This change modifies the local client, where the bypass API is implemented,
to store job data in the driver process until it's polled by the client.
This means no data about a job is transferred back to the client until
the client explicitly asks for it. So in Livy's case, instead of a job
consuming memory in the server process, it would do so in the driver
process, which is owned by the user, and wouldn't affect other users.

Note this change does not implement an API for the data to be retrived
through the Livy server's servlet yet; that will be added separately.