Replace JsonSupport with a Jackson-based servlet.
I don't like json4s; it makes it hard to work with proper types,
and it doesn't understand Java types. This change replaces Scalatra's
JsonSupport, which is based on json4s, with a new 'JsonServlet', which
is based on Jackson.

The new servlet provides routing methods that give the handlers access
to the deserialized type directly, without the need to interact with
Jackson at all; just define the expected type and the servlet handles
the rest.

A corresponding base test class was added to help with writing tests
for these new methods.

As for the existing code, a few changes had to be made:

- The session class hierarchy had to be modified to take a new
  type argument (the Scala/Java type of the request object, so
  that SessionServlet can automatically deserialize it). Methods
  that depended on json4s APIs were modified accordingly.

- The custom serializer for Kind was converted to a Jackson module.

- The json4s Jackson module is still used by the interactive session
  servlet; I added a new unit tests to make sure it's doing what I
  hope is the right thing, but given the previous complete lack of
  tests for that part of the code, it's hard to tell.

In the process, the hacks to be able to use Java types in the client
session backend were completely removed.