Automatically load default configuration from the classpath.
To make it easier to manage client configurations, allow LivyClientBuilder
to load configuration from files present in the classpath, mimicking what
other Hadoop APIs also do. The class will look for two files (livy-client.conf
and spark-defaults.conf) and load their contents if they're present.

On the server side, the ClientSession implementation was changed to
use this instead of trying to load Spark configuration from the system
properties. This should allow server-specific defaults to be configured
while allowing clients to override individual configs using the
session create request.

In the client-local implementation, I broke Spark and Livy configurations
into two separate files, to (arguably) simplify the code a bit. This is
not strictly related to the client configs change; but I also changed the
fact that client-local was loading spark-defaults.conf on its own, and
that's not needed anymore.