Add style checker for copyright header.
Because we have different headers in different modules, some hackery
is needed to pre-process the style checker configuration. And because
maven makes it oh so impossible to figure out the root dir of the
project, this change requires that maven is always executed from the
top-level project directory (otherwise things will fail).