LIVY-61. Simplify LivyConf by extending ClientConf.
That reuses a lot of the type conversion logic; also, declare
some constants in the form of "ConfEntry" objects so that it's
easier to spot their names and default values.

I also changed the server's Main to use a listener instead of
a ScalaBootstrap since that allows using the same LivyConf instance,
instead of having both places initialize the conf object in
different ways.

Finally, I changed the default value for the "environment" config
to "production", since well, the default should not be "development".
The unit tests will set it to "development" explicitly; although it's
all kinda moot since that's not used anywhere.

Closes #52