LIVY-60. Work around Jackson/Scala bug by not using case classes.
Jackson doesn't deal well with Scala case classes; the default values
tend to be ignored. I tried a few workarounds (such as special constructors
annotated with @JsonCreator), but those don't always work (same code
sometimes works and sometimes doesn't - maybe because Scala reflection is
not thread-safe in 2.10?).

So, the less ugly fix was to make the offending types not case classes.
The only affected types seem to be the batch and interactive session
requests, so they were modified to be normal classes with mutable fields.
I added unit tests to make sure the default values are there when the
properties are not provided in the JSON payload.

Closes #54