LIVY-64, LIVY-69. Make JsonServlet process all route outputs.
This change makes JsonServlet map the output object to JSON
whenever the result's type is json, not just when one of the
"j*" methods are used. This means that "get", "post" and friends
also use Jackson to map their results.

As a result, "jget" and "jdelete" are not needed anymore, and
the other methods don't need the weird syntax used to bypass
deserialization of the body. It also fixes existing routes that
were missed when JsonServlet was introduced.