Create a fake SPARK_HOME so tests that fork Spark can run.
This change adds fake "spark-submit" and fake "sparkR" so that tests
can run them as if they were a proper Spark installation. This works
well for Scala / Java, but for python and R it means no pyspark or
SparkR code can actually be tested - just the interaction with the
underlying shell. So a few tests that tried to run Spark code were
removed.