LIVY-66. Allow http client to authenticate using SPNEGO.
The httpclient documentation is kinda off, but this seems to work.
The default value for "RequestConfig.isAuthenticationEnabled" is
actually false (docs say it's "true"), and the SPNEGO auth handler
is installed by default in client instances, so the only thing left
is to register the dummy credentials provider.

Note this requires that the user logs in through some other means
(either using kinit or by performing a JAAS login in their application).