LIVY-73. Allow http client to connect to existing sessions.
This change makes the URI parsing in the HTTP client detect when
the user is trying to reuse an existing session instead of creating
a new one. A new method was added to the client servlet that is
used by the client to ensure the session exists and the user has
permission to use it.

To properly support the functionality, the public API needed to be
changed so that it's possible to clean up a client's state without
shutting down the underlying session.

On slightly related changes, I updated some spots where the code
was failing to update the last activity time of a session, and
changed an internal method name to be more descriptive.