LIVY-76. Fix propagation of Livy config in cluster mode.
Go back to embedding the Livy configuration in the Spark config so
that it's properly propagated to the driver in cluster mode. This
is a little ugly since it requires using Scala APIs in Java code,
but we can't use "--files" and SparkFiles because that class needs
the SparkContext to be initialized, and we need this config before
that.

Also, while testing, I noticed after some debugging that the RPC
code was failing; that's because it was using Spark's Kryo, which
is not binary compatible with the version we compile against. So
now I shade Kryo. I also added a "sync()" call when sending a SASL
message so that if it errors out, we actually get to see the
exception.