LIVY-58. Allow connections between Livy and RSC to be re-established.
This change modifies the direction of connections between Livy and the RSC,
so that it's Livy connecting to the RSC instead of the other way around.
This makes the RSC behave sort of like the current REPL code, where the
context will connect back to Livy to tell its listen address so that Livy
can connect to it.

For that to work, the RemoteDriver now runs an RpcServer, and does not
kill itself when client connections go away. It also allows multiple
concurrent client connections to exist, although it should be rare for
that situation to exist in the real world.

The code to launch Spark was refactored into a separate class, since
when connecting to an existing RSC it needs to be bypassed altogether.