Fix hadoop dependencies scope.
d8842102dd upgraded hadoop dependencies to be "compile" by default,
instead of inheriting the scope from the transitive dependencies.
So the client-http artifact ballooned up to 70MB, and hadoop jars
were now being copied to places they shouldn't.

So override the scope back to "test" or "provided" as needed,
bringing the jar back down to a more manageable 3MB (which is also
much faster to build).