Make re-connections work better.
The previous code had some issues with re-connecting to a remote RSC.
Mainly, it relied on state that might go away if the original client
instance was not around anymore.

So I modified things to that clients are more independent of each other,
and don't rely so much on the ephemeral RpcServer instance that might
exist while waiting for the callback. Process management is also enhanced
a bit, allowing clients to disconnect from the child processes and stop
reading their streams. It remains to be tested whether doing that is safe
and won't lead to the child process deadlocking, since no one is reading
their stdout / stderr anymore.

The new test was modified to stop the original client before trying to
create the new one, which should trigger the situations above.