LIVY-100. Make jars uploaded by the user available in RSC's driver.
Spark's "addJar" only applies to executors, so we need more code to get
jars uploaded by users using the RSC API available to the driver. This
change sets up a new class loader that is used by the "add jar" job to
add uploaded files to the driver's class path.

Closes #99