LIVY-103. Make repl session see user jars, correctly report idle state.
Because of how Spark's REPL deals with class loaders, it's necessary to
poke inside of it to make jars distributed by Livy be seen. This fix is
the same as in the patch below:
  https://github.com/hdinsight/livy/commit/f11cdcb5357046fe555a4ee53c85fb1bccf646b5

Aside from that, the REPL sessions were reporting "idle" state too early.
I had to shuffle a little bit of code around so that the REPL can execute
basic RSC commands, so that the client side can know when it's safe to
send commands. This is a little bit hacky at the moment until we can spend
more time working on LIVY-93.