LIVY-111. Remove async job monitoring API.
This API was a lot more useful when the RSC collected metrics
about running jobs. Since it doesn't do that anymore, this API
is kinda out of place, and there's a lot of hacky code to support
it, so let's just remove it.

Because the HTTP-based client is poll-based, there's no benefit
in having these callback-based APIs. Users can implement the same
functionality by collecting the data they need themselves, and
submitting a job to retrieve that data when desired.

Closes #103