LIVY-108. Remove Guava dependency from RSC.
This avoids two sets of issues: the first is the need to localize the
Guava jar every time an app is started, in case the jars haven't been
cached in HDFS. The second is that we can't really know which version
of Guava will be used; most of the time it will be Hadoop's, but user
configuration and dependencies might change that, and since Guava is
a very common dependency, let's not make user's lives more difficult.

Guava is not explicitly excluded from dependencies, because it's needed
to run the tests (and in some cases is even provided by different
transitive dependencies, such as hive-exec). The patch was tested with
the exclusions in place, to make sure the module compiles without Guava.