[WIP] Add local mini cluster to integration test.
Use MiniDFSCluster and MiniYARNCluster in tests, and refactor some
code so that it's easier to bring up a LivyServer programatically.
This is not yet fully tested (in fact tests downright not work),
but I'll leave the cleanup for a separate change.

TODO: remove the HA-related tests since that code hasn't been merged.