[WIP] Allow cluster config to be defined by user.
Have a default config that starts a small mini cluster, but allow the
config file to be set by the user. This could allow multiple invocations
of the test target in the same pom, or a script that just runs the
integration test module several times with different configs.

The config is also propagated to the cluster implementations (and to the
child processes of the mini cluster).

I also changed the pom to run the tests as part of the regular "test"
target; there's no separate test for that module, so tying this to the
integration test phase is not really needed.