LIVY-123. Initial framework for integration tests.
This change adds an initial framework for integration tests, based on the code
from https://github.com/hdinsight/livy/tree/int-test/.

It allows both real clusters and "mini" clusters (managed by the framework) to be
used. The default is to run a small mini cluster. The code also allows defining your
own cluster config file without having to modify files in the repo for that; that could be
used to run different invocations of the tests as part of the build later on, for example.

There's probably code that needs to be cleaned up or fixed up, but this provides
the basic functionality, at least for mini clusters. I haven't (or hope I haven't) modified
the real cluster code so it should be working as well as it was in the original branch.

Two blatant shortcomings of the current mini cluster is the inability to run pyspark
or SparkR tests. Because it uses a fake Spark installation, it doesn't have access
to the needed python / R libraries. That will be taken care of in separate changes.

Closes #114