LIVY-126. Add integration tests for batches.
Run a couple of simple apps using the batch API. Some adjustment was
needed to make things work; mainly, "spark.submit.deployMode" is not
supported by the default Spark dependency (1.5.0 from CDH), so we need
to use the old style "yarn-cluster" otherwise batches will fail to
launch.

For the tests to work I also needed explicit access to the HDFS cluster,
which means I broke RealCluster in the process. Need to figure out how
to implement that kind of functionality when not using a mini cluster.

Closes #119