LIVY-81. Monitor start of remote context asynchronously.
The remote context can take a while to start, and doing it
synchronously in the builder call makes it possible, even
likely, that you'll get ugly timeouts from the Livy server,
and cause contexts to stay around and be lost while still
consuming resources.

Instead, just start the submission process synchronously, but
wait for the context to connect back asynchronously. This makes
the client create call return much more quickly, but, as with
anything done asynchronously, requires a lot of code to be moved
around and placed in listeners, and to use promises instead of
actual instances in a few places.

The first tricky change was to support submitting jobs before
the context was up. This was done by chaining listeners so that
the actual submission only happens when the context's RPC channel
is actually created; for clients of the API, nothing changes.

The trickier part was monitoring context shut down; this mostly
affected unit tests, since you can't run two Spark contexts in
the same VM (so tests need to make sure the contexts are really
stopped when the stop call returns). That is now done by waiting
for the RPC socket to close, instead of explicitly waiting for
the context thread or process to go away.

As part of the above change I also removed Rpc.Listener since it
can be implemented by adding a listener to the channel's close
future.

Another unrelated thing I ran into while testing was that jobs
can jump directly from SENT to STARTED, and tests were expecting
them to go through QUEUED first. I fixed the test and wrote a
clarification in the API docs.

Closes #120