LIVY-133. Remove default master from interactive session config.
This makes it so that admins have to provide a proper Spark config
to start things up in yarn cluster mode, which is really the only
recommended way to run things.

I also removed setting the master from tests that allow that (those
that run through spark-submit); tests that create a SparkContext
in-process still need to define the master manually.

While testing I found out that the code that caused RSCClient initialization
to fail fast was not working anymore, so I fixed that too. Also, found a
race in the integration tests, where an assert might fail because a session
from a previous test was still in the process of shutting down.