LIVY-129. Add tests for errors in interactive sessions.
The largest portion of this change is a slight refactoring of
how statements are tested in InteractiveIT, to make it easier to
test for errors. With that done, I added a few tests to check error
statement in Scala and Python, and also a test to make sure that
Livy behaves properly when the Spark context dies unexpectedly.

That last one required a fix in the code that handles interactive
sessions, so that errors are properly propagated.

Closes #127