LIVY-31. Close sessions when server shuts down.
Without HA or recovery, not doing this means leaving orphan sessions
that will just hog resources and cannot be used, requiring people to
use YARN's (or other) tools to shut them down.

This change adds a shutdown hook that stops all the active sessions
when the server shuts down. This requires doing more work in a shutdown
hook than is generally palatable, but the alternative is to have some
explicit command (via REST API? monitoring local file? something else?)
to shut down the server, which we don't have.