LIVY-149. Add idle timeout to RSC server.
The RSC server currently listens indefinitely waiting for a Livy server
to connect to it. That can result in sessions that never go away and need
to be manually cleaned up when the Livy server crashes or is otherwise
uncleanly shut down.

So add a timeout, default 10 minutes, that allows sessions to shut themselves
down if the Livy server goes away and doesn't come back.

Closes #137