Make sure original context class loader is restored after calling REPL.
Anytime the Spark REPL is called, it may mess with the thread's context
class loader. There was code to restore the original one after the
interpreter is stopped, but that only applied to a single thread, which
might not even be the right one.

The issue showed up mainly during unit tests, where multiple REPL instances
were created in the same VM. The REPL's class loader would remain as the
active context class loader in many threads in the global executor, and
that would prevent the GC from collecting a whole lot of unused memory
referenced from the REPL.

With this patch, the heap size after a full GC remains constant at around
300M while the repl unit tests are running (while it would balloon to more
than 1G before).

I also added some explicit shut down code to the R interpreter, to make sure
that its backend thread is cleaned up.