LIVY-94. SQLContext/HiveContext is not available in Interpreter
It is straightforward to add SQLContext. But for HiveContext is a little complicated, because it require hive-site.xml and datanucleus jars. There's 2 ways to specify that
1. Ask the client to specify that in rest api, but client should have no idea where the hive-site.xml and datanucleus jars are located in the livy server machine. Although they put these files on hdfs and specify them using hdfs path. But this way still have one issue in yarn-cluster mode due to SPARK-14845
2. Add a option like enableHiveContext in rest api and livy server pick up the hive-site.xml and datanucleus jars automatically. But this way has limitation that it always use the same hive-site.xml, prevent user to user another custom hive-site.xml if he want to connect to another hive metastore.
In this PR, I will combine option 1 and option 2. First check hive-site.ml in user request and then look for it in classpath. And will always detect datanucleus jars automatically.

Closes #132